ncurses 6.1 - patch 20190121
+ add a check in test/configure to work around non-ncurses termcap.h
  file in Slackware.
+ corrected flag for "seq" method of db 1.8.5 interface, needed by toe
  on some of the BSDs.
+ updated "string-hacks" feature.
+ minor improvements to manpage typography.
+ corrected conditionally-compiled limit on color pairs (report by
  "Hudd").
+ add -x option to test/pair_content, test/color_content for testing
  init_extended_pair, extended_pair_content, init_extended_color,
  extended_color_content
+ add -p option to test/pair_content, test/color_content to show the
  return values from the tested functions.
+ improve manual page curs_color.3x discussion of error returns and
  extensions.
+ add O_INPUT_FIELD extension to form library (patch by Leon Winter).
+ override/suppress --enable-db-install if --disable-database configure
  option was given.
+ change a too-large terminal entry in tic from a fatal error to a
  warning (prompted by discussion with Gabriele Balducci).

+++ test/padview.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright (c) 2017 Free Software Foundation, Inc.                        *
+ * Copyright (c) 2017,2019 Free Software Foundation, Inc.                   *
  *                                                                          *
  * Permission is hereby granted, free of charge, to any person obtaining a  *
  * copy of this software and associated documentation files (the            *
@@ -28,7 +28,7 @@
 /*
  * clone of view.c, using pads
  *
- * $Id: padview.c,v 1.12 2017/10/23 00:37:21 tom Exp $
+ * $Id: padview.c,v 1.13 2019/01/21 19:47:07 tom Exp $
  */
 
 #include <test.priv.h>
@@ -89,7 +89,7 @@ show_all(const char *tag, WINDOW *my_pad, int my_row)
     _nc_SPRINTF(temp, _nc_SLIMIT(sizeof(temp))
 		"view %.*s", (int) strlen(tag), tag);
     i = (int) strlen(temp);
-    _nc_SPRINTF(temp + i, _nc_SLIMIT(sizeof(temp) - i)
+    _nc_SPRINTF(temp + i, _nc_SLIMIT(sizeof(temp) - (size_t) i)
 		" %.*s", (int) sizeof(temp) - i - 2, fname);
     mvprintw(0, 0, "%.*s", COLS, temp);
     this_time = time((time_t *) 0);
